requirements.txt
pyautogui
pillow
google-adk
mlx-use